requirements.txt
sanic==18.12.0
aiohttp==3.5.4
aiomysql==0.0.20
aioredis==1.2.0
sqlalchemy==1.3.1
pydub==0.23.0
gunicorn==19.9.0
msgpack
python-dotenv
cryptography==3.3.2 